Added several QOL improvements to the infer_vae.py script.

+++ infer_vae.py
@@ -1,9 +1,5 @@
-import argparse
-import glob
-import hashlib
-import os
-import random
-import re
+import argparse, glob, hashlib
+import os, random, re, shutil
 from dataclasses import dataclass
 from datetime import datetime
 from typing import Optional
@@ -185,6 +181,11 @@ parser.add_argument(
     help="GPU to use in case we want to use a specific GPU for inference.",
 )
 parser.add_argument(
+    "--cpu",
+    action="store_true",
+    help="Use the CPU instead of the GPU, this will be really slow but can be useful for testing or if you dont have a good GPU.",
+)
+parser.add_argument(
     "--max_retries",
     type=int,
     default=30,
@@ -199,6 +200,11 @@ parser.add_argument(
     "--use_paintmind",
     action="store_true",
     help="Use PaintMind VAE..",
+)
+parser.add_argument(
+    "--save_originals",
+    action="store_true",
+    help="Save the original input.png and output.png images to a subfolder instead of deleting them after the grid is made.",
 )
 
 
@@ -356,7 +362,7 @@ def main():
             channels=args.channels,
             layers=args.layers,
             discr_layers=args.discr_layers,
-        ).to(accelerator.device if args.gpu == 0 else f"cuda:{args.gpu}")
+        ).to('cpu' if args.cpu else accelerator.device if args.gpu == 0 else f"cuda:{args.gpu}")
 
         if args.latest_checkpoint:
             accelerator.print("Finding latest checkpoint...")
@@ -421,7 +427,7 @@ def main():
         args.seq_len = vae.get_encoded_fmap_size(args.image_size) ** 2
 
     # move vae to device
-    vae = vae.to(accelerator.device if args.gpu == 0 else f"cuda:{args.gpu}")
+    vae = vae.to('cpu' if args.cpu else accelerator.device if args.gpu == 0 else f"cuda:{args.gpu}")
 
     # Use the parameters() method to get an iterator over all the learnable parameters of the model
     total_params = sum(p.numel() for p in vae.parameters())
@@ -452,7 +458,7 @@ def main():
         )
 
         _, ids, _ = vae.encode(
-            dataset[image_id][None].to(accelerator.device if args.gpu == 0 else f"cuda:{args.gpu}")
+            dataset[image_id][None].to('cpu' if args.cpu else accelerator.device if args.gpu == 0 else f"cuda:{args.gpu}")
         )
         recon = vae.decode_from_ids(ids)
         save_image(recon, f"{args.results_dir}/outputs/output.{str(args.input_image).split('.')[-1]}")
@@ -465,7 +471,7 @@ def main():
         save_image(dataset[image_id], f"{args.results_dir}/outputs/input.png")
 
         _, ids, _ = vae.encode(
-            dataset[image_id][None].to(accelerator.device if args.gpu == 0 else f"cuda:{args.gpu}")
+            dataset[image_id][None].to('cpu' if args.cpu else accelerator.device if args.gpu == 0 else f"cuda:{args.gpu}")
         )
         recon = vae.decode_from_ids(ids)
         save_image(recon, f"{args.results_dir}/outputs/output.png")
@@ -484,7 +490,7 @@ def main():
                     if not args.use_paintmind:
                         # encode
                         _, ids, _ = vae.encode(
-                            dataset[i][None].to(accelerator.device if args.gpu == 0 else f"cuda:{args.gpu}")
+                            dataset[i][None].to('cpu' if args.cpu else accelerator.device if args.gpu == 0 else f"cuda:{args.gpu}")
                         )
                         # decode
                         recon = vae.decode_from_ids(ids)
@@ -493,7 +499,7 @@ def main():
                     else:
                         # encode
                         encoded, _, _ = vae.encode(
-                            dataset[i][None].to(accelerator.device if args.gpu == 0 else f"cuda:{args.gpu}")
+                            dataset[i][None].to('cpu' if args.cpu else accelerator.device if args.gpu == 0 else f"cuda:{args.gpu}")
                         )
 
                         # decode
@@ -520,9 +526,15 @@ def main():
                     filename = f"{hash}_{now}-{os.path.basename(args.vae_path)}.png"
                     grid_image.save(f"{output_dir}/{filename}", format="PNG")
 
-                    # Remove input and output images after the grid was made.
-                    os.remove(f"{output_dir}/input.png")
-                    os.remove(f"{output_dir}/output.png")
+                    if not args.save_originals:
+                        # Remove input and output images after the grid was made.
+                        os.remove(f"{output_dir}/input.png")
+                        os.remove(f"{output_dir}/output.png")
+                    else:
+                        os.makedirs(os.path.join(output_dir, 'originals'), exist_ok=True)
+                        shutil.move(f"{output_dir}/input.png", f"{os.path.join(output_dir, 'originals')}/input_{now}.png")
+                        shutil.move(f"{output_dir}/output.png", f"{os.path.join(output_dir, 'originals')}/output_{now}.png")
+
 
                     del _
                     del ids
